+++ src/llaisys/runtime/output/output.hpp
@@ -22,9 +22,9 @@ public:
     // row: pointer to source logits row.
     // dtype: source row dtype (will be converted to f32).
     // output_id: token index inside current batch.
-    void append_row(const std::byte *row, llaisysDataType_t dtype, int32_t output_id);
+    void append_row(const std::byte *row, llaisysDataType_t dtype, int64_t output_id);
     // Append one output token index without materializing logits.
-    void append_output_id(int32_t output_id);
+    void append_output_id(int64_t output_id);
 
     // Return pointer to the first logits row, or nullptr when empty.
     float *logits() noexcept;
@@ -33,11 +33,11 @@ public:
     // Return number of buffered logits rows.
     int32_t n_outputs() const noexcept;
     // Return row->token-index mapping buffer, or nullptr when empty.
-    const int32_t *output_ids() const noexcept;
+    const int64_t *output_ids() const noexcept;
     // Append one sampled token id for the latest decode row.
-    void append_sampled_id(int32_t sampled_id);
+    void append_sampled_id(int64_t sampled_id);
     // Return sampled token ids aligned with output_ids(), or nullptr when empty.
-    const int32_t *sampled_ids() const noexcept;
+    const int64_t *sampled_ids() const noexcept;
 
 private:
     // Vocabulary size (number of logits per row).
@@ -45,9 +45,9 @@ private:
     // Contiguous logits buffer in row-major layout [n_outputs, voc_].
     std::vector<float> logits_f32_;
     // output_ids_[j] is the batch token index corresponding to row j.
-    std::vector<int32_t> output_ids_;
+    std::vector<int64_t> output_ids_;
     // sampled_ids_[j] is sampled token id for output row j.
-    std::vector<int32_t> sampled_ids_;
+    std::vector<int64_t> sampled_ids_;
 };
 
 } // namespace llaisys::runtime::output
